Add modular features, model fitting and evaluation with tests
