+++ scripts/detect_regime.py
@@ -21,7 +21,11 @@ try:  # optional dependency
 except Exception:  # pragma: no cover - optional
     hdbscan = None
 
-from train_target_clone import _load_logs, _extract_features, _load_calendar  # type: ignore
+from train_target_clone import _load_logs, _load_calendar  # type: ignore
+try:
+    from scripts.features import _extract_features  # type: ignore
+except Exception:
+    from features import _extract_features  # type: ignore
 
 
 def detect_regimes(
