+++ tests/test_evaluate.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import sys
 
 sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-from scripts.evaluate_predictions import evaluate
+from scripts.evaluation import evaluate
 
 
 def _write_prediction(file: Path, ts: str, direction: str = "buy"):
